edgeyolo-main/edgeyolo-main/requirements.txt
torch
torchvision
opencv-python
numpy
loguru
pycocotools
requests
tqdm
tabulate
thop
pyyaml
pandas
pathlib
pillow
matplotlib
mahotas
seaborn
scipy
onnx
onnxsim
mahotas
###### tensorrt export #####
# pycuda
# tensorrt
# torch2trt
